+++ packages/bluesky/lib/src/entities/actor.dart
@@ -33,6 +33,14 @@ class Actor with _$Actor {
 
   factory Actor.fromJson(Map<String, Object?> json) => _$ActorFromJson(json);
 
+  /// Returns true if authenticated user has not muted yet this actor,
+  /// otherwise false.
+  bool get isNotMuted => viewer.isNotMuted;
+
+  /// Returns true if authenticated user has not blocked yet by this actor,
+  /// otherwise false.
+  bool get isNotBlockedBy => viewer.isNotBlockedBy;
+
   /// Returns true if authenticated user has already muted this actor by list,
   /// otherwise false.
   bool get isMutedByList => viewer.isMutedByList;