+++ packages/bluesky/lib/src/entities/list_viewer.dart
@@ -12,6 +12,9 @@ part 'list_viewer.g.dart';
 
 @freezed
 class ListViewer with _$ListViewer {
+  // ignore: unused_element
+  const ListViewer._();
+
   @JsonSerializable(includeIfNull: false)
   const factory ListViewer({
     @JsonKey(name: 'muted') required bool isMuted,
@@ -19,4 +22,8 @@ class ListViewer with _$ListViewer {
 
   factory ListViewer.fromJson(Map<String, Object?> json) =>
       _$ListViewerFromJson(json);
+
+  /// Returns true if authenticated user has not muted yet this actor,
+  /// otherwise false.
+  bool get isNotMuted => !isMuted;
 }